refactor(data): split finetuning batch utilities
